Prefer to keep all P2P utils into a single folder
